+++ src/main/java/team/themoment/hellogsmv3/domain/application/entity/abs/AbstractMiddleSchoolGrade.java
@@ -11,7 +11,6 @@ import java.util.UUID;
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Inheritance(strategy = InheritanceType.SINGLE_TABLE)
-@ToString
 public abstract class AbstractMiddleSchoolGrade {
 
     @Id
