+++ src/main/java/team/themoment/hellogsmv3/domain/application/entity/abs/AbstractPersonalInformation.java
@@ -12,7 +12,6 @@ import java.util.UUID;
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Inheritance(strategy = InheritanceType.SINGLE_TABLE)
-@ToString
 public abstract class AbstractPersonalInformation {
 
     @Id
